simplify table creation process
